Change tokensRequested datatype (#6)
Change `tokensRequested` data type to STRING. Addresses #2.

Previously, `tokensRequested` had the datatype INTEGER. This was
problematic because JS cannot handle integers larger than 53 bits,
and the EVM only supports integers. Indeed, if a user requested a
a fractional amount of tokens, JS may not have been able to convert
the value with precision. This commit addresses this issue by changing
the datatype to STRING.

* update proposalSchema, model, and data in API
  * add migration
  * add tests
* change datatype to string in client
* add custom validation for token amount
* convert and correctly display tokensRequested on client side