Update deployments (#16)
This work changes the way we're passing configuration to the applications. It adds a new Helm chart `panvala-base`, that sets up configuration that the other applications can then read from. This configuration includes a reference to the database, the database configuration, and the addresses of the contracts. Instead of defining these values in the application deployments, they now read from the configuration that has already been deployed. In addition, it contains some improvements to make it easier to debug failures and misconfigurations.

* Add `panvala-base` Helm chart
* Update `panvala-api` and `panvala-frontend` charts to use resources deployed by `panvala-base`
* Add new scripts for deploying api and frontend, and update CI deploy to use them
* Add readiness probe to API at `/ready` - check database and web3 connections
* Add HTTP logging to API
* Add Makefile