fix delegated commitBallot voting:
previously, the determination of `numTokens` was taking up space and
execution of `commitBallot` did not include the new `tokenHolder`
address parameter. additionally, the hotfix that was made last week did
not take into account that `numTokens` was wrong, as it was directly
calculated using the currently signed in account. this commit
moves some delegation logic to where numTokens is correctly determined.