Fix generation of commit message and hash
There were inconsistencies in the way the commit message and
commit hash were being generated and validated. Also, in the
previous implementation, the order of the keys in the vote
choices objects mattered when calculating these values, leading
to failures in validation (and in some cases reveals, depending
on the JS implementation).

* Fix generation of commit message and commit hash to be order-
  independent (resources are processed in alphabetical order)
* Fix validation of commit message on the API - it was iterating
  through choices instead of passing all of them to
  generateCommitMessage()
* Fix generation of commit message on vote page to include all
  choices - it was only sending for grants
* Clean up formatting of commit message
* Update API tests