Prevent uninitialized paramstore (#156)
Address audit issue 7.9.

the owner of ParameterStore may choose to never call init() and remain in control of the ParameterStore being able to arbitrarily change values without requiring a proposal to perform it (via setInitialValue()). this pr prevents calls to createProposal and setValue if the value of initialized is false. this pr also adds a public visibility modifier to the initialized value so as to be transparent about the state of the contract