Show a message for categories when submission closes
At /slates/create/* show a message with a link to the home page and the
slate submission close date for the given category if it is closed.
Otherwise, show the form.

Set explicit page status in /slates/create/* to fix the issue where a
message about the expiration was being shown before the ballot was
loaded properly.

- Add CloseSlateSubmission component and show when submission is closed
- Show "Loading..." string until ballot data loads
- Add field slateSubmissionStart date to IBallotDates