Improve governance slate creation
Make the governance slate metadata match the grant slate metadata.
Properly set and save the `oldValue` for parameter changes in the
governance slate form. Make it easier to detect duplicate values by
storing the raw old value in addition to the display value in the
form.

Return the transformed old governance proposals in getAllSlates()
with a key `parameterChanges`, to match the expected display.

IPFS fixes
- Use asynchronous IPFS add (saveToIpfs())
- Fix IPFS add utility